Create stack04.rb
First, we read the number of videos on the shelf, n, from standard input and create an empty array called shelf to represent the shelf.
We loop n times and read the control numbers of the videos from standard input, adding them to the shelf array.
Next, we read the number of borrowing/returning operations, k, from standard input.
We loop k times and read each operation from standard input. If the operation is a rent, we simply remove the top video from the shelf array using the pop method. If the operation is a return, we extract the control number of the video from the operation string, convert it to an integer, and add it to the shelf array using the << (append) method.
Finally, we output the contents of the shelf array in reverse order, since the videos are arranged in the opposite order from how they were input.
Note that we don't need to use a stack data structure explicitly, as Ruby's array provides all the necessary stack operations.